backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-telegram-bot
python-dotenv
python-multipart
pydantic
pydantic-settings
typing_extensions>=4.5.0
